Modify ball behavior
Ball increases in vertical speed each time it hits the paddle up to a maximum speed, where it then remains constant until the ball is lost. This is to increase difficulty and make clearing levels faster.

+++ src/main/java/com/hcyic1/brickdestroy/view/DebugConsole.java
@@ -97,7 +97,7 @@ public class DebugConsole extends JDialog implements WindowListener {
     public void windowActivated(WindowEvent windowEvent) {
         setLocation();
         Ball b = level.ball;
-        debugPanel.setSpeeds(b.getSpeedX(), b.getSpeedY());
+        debugPanel.setSpeeds((int) b.getSpeedX(), (int) b.getSpeedY());
     }
 
     @Override
